pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stock-monitor"
version = "1.0.0"
description = "Advanced Stock Market Monitoring and Analytics System"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "CuteSurtr"}
]
maintainers = [
    {name = "CuteSurtr"}
]
keywords = ["stock market", "monitoring", "analytics", "portfolio", "trading", "finance"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Office/Business :: Financial :: Investment",
    "Topic :: Scientific/Engineering :: Information Analysis",
]
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "sqlalchemy>=2.0.23",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.9",
    "asyncpg>=0.29.0",
    "influxdb-client>=1.38.0",
    "yfinance>=0.2.28",
    "alpha-vantage>=2.3.1",
    "requests>=2.31.0",
    "websockets>=12.0",
    "aiohttp>=3.9.1",
    "pandas>=1.5.0,<2.1.0",
    "numpy>=1.21.0,<1.25.0",
    "scikit-learn>=1.0.0,<1.4.0",
    "scipy>=1.7.0,<1.12.0",
    "ta>=0.10.2",
    "statsmodels>=0.13.0,<0.14.0",
    # "tensorflow>=2.15.0",  # Commented out - not compatible with Python 3.13
    # "torch>=1.7.0",        # Commented out - not compatible with Python 3.13
    "plotly>=5.24.1",
    "dash>=2.17.1",
    "dash-bootstrap-components>=1.5.0",
    "matplotlib>=3.5.0,<3.8.0",
    "seaborn>=0.11.0,<0.13.0",
    "prometheus-client>=0.19.0",
    "prometheus-fastapi-instrumentator>=6.1.0",
    "psutil>=5.9.6",
    "newspaper3k>=0.2.8",
    "textblob>=0.17.1",
    "nltk>=3.8.1",
    "beautifulsoup4>=4.12.2",
    "redis>=4.6.0",
    "celery>=5.3.4",
    "flower>=2.0.1",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
    "schedule>=1.2.0",
    "click>=8.1.7",
    "rich>=13.7.0",
    "pytz>=2023.3",
    "python-dateutil>=2.8.2",
    "ratelimit>=2.2.1",
    "structlog>=23.2.0",
    "cryptography>=43.0.0",
    "passlib>=1.7.4",
    "selenium>=4.15.2",
    "webdriver-manager>=4.0.1",
    "investpy>=1.0.8",
    "cvxpy>=1.4.1",
    "pyportfolioopt>=1.5.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "black>=23.11.0",
    "flake8>=6.1.0",
    "mypy>=1.7.1",
]
desktop = [
    "psutil>=5.9.6",
    "rich>=13.7.0",
]

[project.scripts]
stock-monitor = "src.main:main"
stock-monitor-worker = "src.celery_app:celery_app"

[project.urls]
Homepage = "https://github.com/CuteSurtr/monitor-advanced"
Documentation = "https://github.com/CuteSurtr/monitor-advanced/tree/main/docs"
Repository = "https://github.com/CuteSurtr/monitor-advanced"
"Bug Tracker" = "https://github.com/CuteSurtr/monitor-advanced/issues"

[tool.setuptools]
packages = ["src"]

[tool.setuptools.package-data]
"*" = ["*.txt", "*.md", "*.yaml", "*.yml", "*.json"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"
pythonpath = ["."]

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "yfinance.*",
    "alpha_vantage.*",
    "newspaper3k.*",
    "investpy.*",
    "cvxpy.*",
    "pyportfolioopt.*",
]
ignore_missing_imports = true
